+++ src/main/kotlin/de/achimonline/quickjinja/python/QuickJinjaPython.kt
@@ -8,6 +8,8 @@ class QuickJinjaPython {
         fun createScriptFile(
             templateFile: File,
             variablesFile: File? = null,
+            customFiltersFile: File? = null,
+            customTestsFile: File? = null,
             options: Map<String, Any>? = null,
             fileSystemLoaderPaths: List<String> = emptyList()
         ): File {
@@ -15,6 +17,14 @@ class QuickJinjaPython {
                 import sys
                 import json
                 import jinja2
+                import importlib.util
+
+                def import_from_path(module_name, file_path):
+                    spec = importlib.util.spec_from_file_location(module_name, file_path)
+                    module = importlib.util.module_from_spec(spec)
+                    sys.modules[module_name] = module
+                    spec.loader.exec_module(module)
+                    return module
                 
                 options = {
             """.trimIndent()
@@ -51,26 +61,49 @@ class QuickJinjaPython {
                     environment.tests.update(ansible.plugins.test.core.TestModule().tests())
                 except:
                     ...
-
-                with open(r'${templateFile.absolutePath}') as template_file:
-                    template = template_file.read()
-
             """.trimIndent()
 
-            scriptFileContent += if (variablesFile != null) {
-                """${System.lineSeparator()}
-                    with open(r'${variablesFile.absolutePath}') as data_file:
-                        data = json.load(data_file)
+            if (customFiltersFile != null && customFiltersFile.exists()) {
+                scriptFileContent += """${System.lineSeparator()}
 
+                    try:
+                        environment.filters.update(import_from_path('quick_jinja_custom_filters', '${customFiltersFile.absolutePath}').filters())
+                    except:
+                        ...
                 """.trimIndent()
-            } else {
-                """${System.lineSeparator()}
-                    data = {}
+            }
 
+            if (customTestsFile != null && customTestsFile.exists()) {
+                scriptFileContent += """${System.lineSeparator()}
+
+                    try:
+                        environment.tests.update(import_from_path('quick_jinja_custom_tests', '${customTestsFile.absolutePath}').tests())
+                    except:
+                        ...
                 """.trimIndent()
             }
 
             scriptFileContent += """${System.lineSeparator()}
+
+                with open(r'${templateFile.absolutePath}') as template_file:
+                    template = template_file.read()
+            """.trimIndent()
+
+            scriptFileContent += if (variablesFile != null) {
+                """${System.lineSeparator()}
+
+                    with open(r'${variablesFile.absolutePath}') as data_file:
+                        data = json.load(data_file)
+                """.trimIndent()
+            } else {
+                """${System.lineSeparator()}
+
+                    data = {}
+                """.trimIndent()
+            }
+
+            scriptFileContent += """${System.lineSeparator()}
+
                 try:
                     print(environment.from_string(template).render(**data))
                 except Exception as exception: